cartpole/backend/requirement.txt
fastapi==0.95.0
gym==0.26.2
numpy==1.24.2
Pillow==9.5.0
uvicorn==0.22.0